+++ src/main/scala/diplomaticobjectmodel/model/CustomISAExtensions.scala
@@ -1,5 +1,21 @@
 package diplomaticobjectmodel.model
 
-class CustomISAExtensions {
+import freechips.rocketchip.diplomaticobjectmodel.model.{OMEnum, Xsifivecflushdlone}
 
+trait OMCustomExtensionSpecification{
+  def name: String
+  def version: String
+  def _types: Seq[String] = Seq("OMCustomExtensionSpecification", "OMSpecification")
+}
+
+object CustomISAExtensions {
+
+  trait OMCustomExtensionType extends OMEnum
+  case object XsifivecflushdloneKey extends OMCustomExtensionType
+
+  val cflush: (String) => OMCustomExtensionSpecification = (s: String) => Xsifivecflushdlone("SiFive Extension for Cache Flush", s)
+
+  val customSpecifications = Map[OMCustomExtensionType, (String) => OMCustomExtensionSpecification](
+    XsifivecflushdloneKey -> cflush
+  )
 }
