+++ src/main/scala/diplomaticobjectmodel/model/ISASpecifications.scala
@@ -9,6 +9,7 @@ sealed trait PrivilegedArchitectureExtension extends OMEnum
 case object MachineLevelISA extends PrivilegedArchitectureExtension
 case object SupervisorLevelISA extends PrivilegedArchitectureExtension
 
+
 object PrivilegedArchitectureExtensions {
   val specifications = Map[PrivilegedArchitectureExtension, String](
     MachineLevelISA -> "Machine-Level ISA",
@@ -40,11 +41,6 @@ object ISAExtensions {
   )
 
   def specVersion(extension: OMExtensionType, version: String): OMSpecification = OMSpecification(specifications(extension), version)
-
-  def customExtensions(coreParams: CoreParams): List[OMCustomExtensionSpecification] = {
-    val cflush = coreParams.haveCFlush.option(Xsifivecflushdlone("", ""))
-    List[Option[OMCustomExtensionSpecification]](cflush).flatten
-  }
 }
