+++ src/main/scala/diplomaticobjectmodel/model/OMISA.scala
@@ -2,7 +2,10 @@
 
 package freechips.rocketchip.diplomaticobjectmodel.model
 
+import diplomaticobjectmodel.model.CustomISAExtensions.XsifivecflushdloneKey
+import diplomaticobjectmodel.model.{CustomISAExtensions, OMCustomExtensionSpecification}
 import freechips.rocketchip.rocket.RocketCoreParams
+import freechips.rocketchip.tile.CoreParams
 import freechips.rocketchip.util.BooleanToAugmentedBoolean
 
 trait OMExtensionType extends OMEnum
@@ -24,6 +27,7 @@ case object RV32E extends OMBaseInstructionSet
 case object RV32I extends OMBaseInstructionSet
 case object RV64I extends OMBaseInstructionSet
 case object RV128I extends OMBaseInstructionSet
+
 
 case class Xsifivecflushdlone(
   name: String,
@@ -48,6 +52,13 @@ case class OMISA(
 ) extends OMCompoundType
 
 object OMISA {
+  def customExtensions(coreParams: CoreParams): List[OMCustomExtensionSpecification] = {
+    val cflush = coreParams.haveCFlush.option(CustomISAExtensions.customSpecifications(XsifivecflushdloneKey)(""))
+    List[Option[OMCustomExtensionSpecification]](
+      cflush
+    ).flatten
+  }
+
   def rocketISA(coreParams: RocketCoreParams, xLen: Int): OMISA = {
     val baseInstructionSet = xLen match {
       case 32 => RV32I
@@ -84,7 +95,7 @@ object OMISA {
       u = (coreParams.useVM || coreParams.useUser).option(isaExtSpec(U, "1.10")),
       s = coreParams.useVM.option(isaExtSpec(S, "1.10")),
       addressTranslationModes = Seq(addressTranslationModes),
-      customExtensions = ISAExtensions.customExtensions(coreParams)
+      customExtensions = customExtensions(coreParams)
     )
   }
 }
